# sqlparser: Keep track of source token positions