--- a/go/vt/sqlparser/Makefile
+++ b/go/vt/sqlparser/Makefile
@@ -15,7 +15,7 @@
 MAKEFLAGS = -s
 
 sql.go: sql.y
-	go run ./goyacc -fast-append -o sql.go sql.y
+	go run ./goyacc -fast-append -token-pos -o sql.go sql.y
 	gofmt -w sql.go
 
 clean: