--- a/go/vt/sqlparser/goyacc/goyacc.go
+++ b/go/vt/sqlparser/goyacc/goyacc.go
@@ -18,6 +18,7 @@ file such as NOTICE, LICENCE or COPYING.
 	Portions Copyright © 2005-2007 C H Forsyth (forsyth@terzarima.net)
 	Revisions Copyright © 2000-2007 Lucent Technologies Inc. and others
 	Portions Copyright © 2009 The Go Authors. All rights reserved.
+	Portions Copyright © 2021 The Vitess Authors.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal