--- a/go/vt/sqlparser/parser.go
+++ b/go/vt/sqlparser/parser.go
@@ -28,36 +28,27 @@ import (
 )
 
 // parserPool is a pool for parser objects.
-var parserPool = sync.Pool{}
+var parserPool = sync.Pool{
+	New: func() interface{} {
+		return &yyParserImpl{}
+	},
+}
 
 // zeroParser is a zero-initialized parser to help reinitialize the parser for pooling.
-var zeroParser = *(yyNewParser().(*yyParserImpl))
+var zeroParser yyParserImpl
 
 // MySQLVersion is the version of MySQL that the parser would emulate
 var MySQLVersion string = "50709"
 
 // yyParsePooled is a wrapper around yyParse that pools the parser objects. There isn't a
-// particularly good reason to use yyParse directly, since it immediately discards its parser.  What
-// would be ideal down the line is to actually pool the stacks themselves rather than the parser
-// objects, as per https://github.com/cznic/goyacc/blob/master/main.go. However, absent an upstream
-// change to goyacc, this is the next best option.
+// particularly good reason to use yyParse directly, since it immediately discards its parser.
 //
 // N.B: Parser pooling means that you CANNOT take references directly to parse stack variables (e.g.
 // $$ = &$4) in sql.y rules. You must instead add an intermediate reference like so:
 //    showCollationFilterOpt := $4
 //    $$ = &Show{Type: string($2), ShowCollationFilterOpt: &showCollationFilterOpt}
 func yyParsePooled(yylex yyLexer) int {
-	// Being very particular about using the base type and not an interface type b/c we depend on
-	// the implementation to know how to reinitialize the parser.
-	var parser *yyParserImpl
-
-	i := parserPool.Get()
-	if i != nil {
-		parser = i.(*yyParserImpl)
-	} else {
-		parser = yyNewParser().(*yyParserImpl)
-	}
-
+	parser := parserPool.Get().(*yyParserImpl)
 	defer func() {
 		*parser = zeroParser
 		parserPool.Put(parser)