--- a/go/vt/sqlparser/sql.y
+++ b/go/vt/sqlparser/sql.y
@@ -25,8 +25,8 @@ func setAllowComments(yylex interface{}, allow bool) {
   yylex.(*Tokenizer).AllowComments = allow
 }
 
-func setDDL(yylex interface{}, node interface{}) {
-  yylex.(*Tokenizer).partialDDL = node.(Statement)
+func setDDL(yylex interface{}, node Statement) {
+  yylex.(*Tokenizer).partialDDL = node
 }
 
 func incNesting(yylex interface{}) bool {