## Description

This is not an actual performance improvement! The horror!

This is a feature/bugfix PR for a use case that @systay mentioned: Sometimes, our users use complex expressions as column names. Our SQL parser normalizes all the SQL syntax, which is useful in most cases, but for column names, it would be helpful to be able to access the column name _exactly as it was in the original query_ (i.e. without any normalization), so it can be used when rewriting queries.

A simple example would be a `SELECT 1+1, 1 + 1` query, which always gets normalized as `SELECT 1+1, 1+1`. The fact that the second column has whitespace between the operators is lost in our current version of the parser.

The fix for this is rather elaborate: it requires propagating metadata on the position for the stard/end tokens in a query. That's expensive! I've been pondering this _a lot_ and it's just very hard to only scope down the tracking of token positions to "column names", because in our grammar an aliased column name can contain any `Expr`, which accounts for like 90% of all the productions. It's just significantly simpler to propagate token positions for the whole grammar, but that comes with a performance penalty.

The approach presented in this PR (commit https://github.com/vitessio/vitess/commit/d99e2d9b9c29efa75aee0b1d1e35fb33aad1bf65 -- everything else is my Yacc codegen refactoring PR which hasn't been merged to `master` yet) modifies the actual codegen to optimize as much as possible the propagation of token positions.

Despite this, it adds enough overhead to be noticeable in benchmarks:

```
name                               old time/op    new time/op    delta
ParseTraces/django_queries.txt-16    1.46ms ± 2%    1.49ms ± 4%  +2.34%  (p=0.029 n=10+10)
ParseTraces/lobsters.sql.gz-16       62.2ms ± 4%    65.0ms ± 3%  +4.50%  (p=0.000 n=10+10)
ParseStress/sql0-16                  7.74µs ± 3%    8.32µs ± 2%  +7.45%  (p=0.000 n=9+10)
ParseStress/sql1-16                  24.4µs ± 4%    25.4µs ± 4%  +3.99%  (p=0.001 n=10+10)
Parse3/normal-16                     1.73ms ± 2%    1.73ms ± 3%    ~     (p=0.905 n=9+10)
Parse3/escaped-16                    6.18ms ± 1%    6.16ms ± 1%    ~     (p=0.393 n=10+10)

name                               old alloc/op   new alloc/op   delta
ParseTraces/django_queries.txt-16     239kB ± 0%     246kB ± 0%  +3.16%  (p=0.000 n=9+10)
ParseTraces/lobsters.sql.gz-16       10.9MB ± 0%    11.1MB ± 0%  +2.03%  (p=0.000 n=10+10)
ParseStress/sql0-16                  1.52kB ± 0%    1.60kB ± 0%  +5.26%  (p=0.000 n=10+10)
ParseStress/sql1-16                  5.00kB ± 0%    5.16kB ± 0%  +3.20%  (p=0.000 n=8+10)
Parse3/normal-16                     3.17kB ± 0%    3.20kB ± 0%  +1.01%  (p=0.000 n=10+10)
Parse3/escaped-16                    5.23MB ± 0%    5.23MB ± 0%    ~     (p=0.516 n=10+10)
````

@systay believes that the ~5% performance hit is not worth it for the functionality this provides. I don't have any opinion on the matter because I don't fully understand what's the actual use case for this feature (lol).

Leaving this here for discussion.

## Related Issue(s)
<!-- List related issues and pull requests: -->

- 

## Checklist
- [ ] Should this PR be backported?
- [x] Tests were added or are not required
- [x] Documentation was added or is not required

## Deployment Notes
<!-- Notes regarding deployment of the contained body of work. These should note any db migrations, etc. -->

## Impacted Areas in Vitess
Components that this PR will affect:

- [x]  Query Serving
- [ ]  VReplication
- [ ]  Cluster Management
- [ ]  Build/CI
- [ ]  VTAdmin