@derekperkins 2021-03-15
> I don't fully understand what's the actual use case for this feature

I don't either, so I'm on Team @systay