@systay 2021-03-15
The use case is to mimic MySQL so closely that we don't surprise users. MySQL uses the exact string behind the expression, and Vitess doesn't. When I first discussed this with @vmg, my assumption was that we were already keeping track of this information so all we needed was to expose it. If it was for free, I think it would be nice to mimic MySQL here, but I don't think it's worth a consistent 5% performance drop for all parsing.