@sougou 2021-03-15
Taking a step back. Beyond providing compatibility, I think there is very little functional usefulness for this behavior. In contrast, the complexity it introduces is astounding. I feel like it's not worth it.