@deepthi 2021-03-15
I agree with @systay, the 5% penalty doesn't seem worth it.