@shlomi-noach 2021-03-16
Late for the party, I agree to the above comments. Developers should really use `AS some_alias` for complex column expressions.